+++ rest_framework_services/types/__init__.py
@@ -12,6 +12,8 @@ from rest_framework_services.types.dispatch_result import DispatchResult
 from rest_framework_services.types.field_change import FieldChange
 from rest_framework_services.types.http_extras import HttpExtras
 from rest_framework_services.types.no_input import NoInput
+from rest_framework_services.types.offline_context import OfflineContext
+from rest_framework_services.types.offline_service_view import OfflineServiceView
 from rest_framework_services.types.selector_kind import SelectorKind
 from rest_framework_services.types.selector_spec import SelectorSpec
 from rest_framework_services.types.service_spec import ServiceSpec
@@ -27,6 +29,8 @@ __all__ = [
     "FieldChange",
     "HttpExtras",
     "NoInput",
+    "OfflineContext",
+    "OfflineServiceView",
     "SelectorKind",
     "SelectorSpec",
     "ServiceSpec",